chore: separate `keys`
