+++ keys/keys.go
@@ -1,4 +1,4 @@
-package transaction
+package keys
 
 import (
 	"errors"
